added authentication

+++ FrontEnd(react)/TodoApp/src/Hooks/useDeleteTodo.tsx
@@ -1,10 +1,20 @@
 import { useState } from "react";
+import { getToken } from "./useLoginUser";
 
 export const useDeleteTodo = () => {
   const [url] = useState("https://localhost:44300/todos");
   const deleteTodo = async (id: number) => {
+    const token = getToken();
+    if (!token) {
+      console.log("No token found");
+      return;
+    }
+
     const response = await fetch(`${url}/${id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     });
 
     if (!response.ok) {
@@ -15,4 +25,3 @@ export const useDeleteTodo = () => {
   };
   return deleteTodo;
 };
-
